+++ packages/core/src/core-providers/structure/structure.types.ts
@@ -1,5 +1,5 @@
-import type { TFieldStructure } from '../../../types/base.types';
-import type { TProviderService } from '../../../types/provider.types';
+import type { TFieldStructure } from '../../types/base.types';
+import type { TProviderService } from '../../types/provider.types';
 type ScreenName = string;
 
 export type EndStructure = Record<string, TFieldStructure>;
