+++ packages/core/src/index.ts
@@ -1,6 +1,6 @@
 export { toRxStore } from './base/store';
 export { FormBuilder } from './base/builder';
-export * as BuiltInProviders from './providers/built-in';
+export * as BuiltInProviders from './core-providers';
 export { ViewGenerator } from './generate';
 
 export * as TBase from './types/base.types';
